# Team-6_A4/openapi.yaml
openapi: 3.0.3
info:
  title: CampusEats Order Service
  version: 1.0.0
  description: REST API for the CampusEats Order Service.

servers:
  - url: http://localhost:5000
    description: Local development server

tags:
  - name: Orders
    description: CampusEats order operations
  - name: Cancellation
    description: Order cancellation operations

paths:
  /orders:
    post:
      tags: [Orders]
      summary: Create an order
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              description: URI of the newly created order
              schema:
                type: string
                example: /orders/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Malformed request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Idempotency key conflicts with a different request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Valid request refused by the order or payment domain
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Payments service remained unavailable after safe retries
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrders
      parameters:
        - name: studentId
          in: query
          required: false
          description: Return only orders belonging to this student.
          schema:
            type: string
            minLength: 1
          example: STU-1001
        - name: status
          in: query
          required: false
          description: Return only orders with this status.
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Matching orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get one order
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /orders/{orderId}/cancellation:
    post:
      tags: [Cancellation]
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order was cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Order cannot be cancelled in its current state
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Cancellation is refused by the domain
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Key that makes order creation safely repeatable.
      schema:
        type: string
        minLength: 1
    OrderId:
      name: orderId
      in: path
      required: true
      description: Public order identifier.
      schema:
        type: integer
        minimum: 1

  schemas:
    CreateOrderRequest:
      type: object
      required:
        - studentId
        - items
        - paymentMethod
      additionalProperties: false
      properties:
        studentId:
          type: string
          minLength: 1
          example: STU-1001
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        paymentMethod:
          type: string
          minLength: 1
          example: CARD

    OrderItemRequest:
      type: object
      required:
        - itemId
        - quantity
      additionalProperties: false
      properties:
        itemId:
          type: string
          minLength: 1
          example: ITEM-101
        quantity:
          type: integer
          minimum: 1
          example: 2
        unitPrice:
          type: number
          format: double
          minimum: 0
          description: Price supplied by the client for this assignment's in-process catalogue representation.
          example: 120.00

    OrderResponse:
      type: object
      required:
        - orderId
        - status
        - total
        - estimatedReadyMinutes
      properties:
        orderId:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/OrderStatus'
        total:
          type: number
          format: double
          example: 240.00
        estimatedReadyMinutes:
          type: integer
          minimum: 0
          example: 15

    OrderStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - CANCELLED
        - COMPLETED

    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: https://campuseats.example.com/problems/order-not-found
        title:
          type: string
          example: Order not found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: No order exists with id 42